# app/server/static/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Search Intelligence API",
    "description": "AI-powered career intelligence platform — job discovery, company analysis, skill management, resume generation, and career insights.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Development server" }
  ],
...
  "paths": {
    "/api/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get paginated list of processed jobs",
        "parameters": [
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 30 } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["created_at", "overall_score", "fit_score", "success_score", "num", "company", "location"] } },
          { "name": "sort_dir", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "filter_tech", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated tech stack filter" },
          { "name": "filter_cities", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated city filter" }
        ],
        "responses": { "200": { "description": "Paginated job list with total count" } }
      }
    },
    "/api/jobs/{num}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get a single job by number",
        "parameters": [{ "name": "num", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Job details" }, "404": { "description": "Job not found" } }
      },
      "put": {
        "tags": ["Jobs"],
        "summary": "Update job fields",
        "parameters": [{ "name": "num", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Updated job" } }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete a job (hard delete)",
        "parameters": [{ "name": "num", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Job deleted" } }
      }
    },
    "/api/jobs/{num}/requeue": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Re-queue a processed job for reprocessing",
        "parameters": [{ "name": "num", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Job re-queued" } }
      }
    },
    "/api/jobs/{num}/rescore": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Rescore an existing job",
        "parameters": [{ "name": "num", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Job rescored" } }
      }
    },
...
  }
}